v1.101013 	Enh.		Operation expression. The first input is the expression, subsequent inputs are named. 	BUG:		The bitprecision was capped at 63, even though the error message says 64.
v1.101011
	Enh.		Operation random. takes 2 values representing the high/low values of the interval.
				so eg. precision 0 with inputs 1 and 2 will return either 1 or 2. 
				But precision 1 with the same inputs will return 1.0,1.1,1.2,1.3, ... 2.0
				The precision attribute affects the number of decimal points (and therefore the size of the interval)
	BUG:		In errors, mysql password was being shown. Now it is obscured with eg: p******d.

+++ commons/environment/environment.h
@@ -64,7 +64,7 @@ private:
 	string gRootDir;
 	string gScriptsDir;
 	string gScratchDir;
-	double basetime;	//used for timing.
+	long double basetime;	//used for timing.
 	std::string empty;
 	std::string parmprefix;
 	//-- The following are RESPONSE cookies
